test gnss connection loss
